requirements.txt
# Simple dependencies for running the scripts
requests>=2.31.0
python-dotenv>=1.0.0

# Optional: For running tests
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
